pyproject.toml
[project]
name = "nopii"
version = "0.1.0"
description = "A batteries-included Python toolkit for detecting, transforming, masking, pseudonymizing, and auditing PII across data engineering workflows"
readme = "README.md"
authors = [
    { name = "ay-mich"}
]
license = "Apache-2.0"
requires-python = ">=3.12"
keywords = ["pii", "privacy", "data-protection", "transformation", "compliance", "gdpr", "ccpa"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "jinja2>=3.0.0",
    "rich>=13.0.0",
    "typing-extensions>=4.5.0",

]

[project.optional-dependencies]
pandas = ["pandas>=2.0.0", "numpy>=1.24.0"]
report-html = ["plotly>=5.0.0", "kaleido>=0.2.1", "weasyprint>=59.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "ruff>=0.0.280",
    "bandit>=1.7.5",
    "vulture>=2.14",
    "twine>=6.2.0",
    "pip-audit>=2.9.0"
]
# keep for compatibility if needed in the future (currently minimal extras)

[project.scripts]
nopii = "nopii.cli:main"
no-pii = "nopii.cli:main"

[project.urls]
Homepage = "https://github.com/ay-mich/nopii"
Documentation = "https://nopii.readthedocs.io"
Repository = "https://github.com/ay-mich/nopii"
Issues = "https://github.com/ay-mich/nopii/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_unused_ignores = true
ignore_missing_imports = true
no_implicit_optional = false
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
# Initial release: non-strict. Tighten incrementally per module.
files = [
  "src/nopii/core/**/*.py",
  "src/nopii/detectors/**/*.py",
  "src/nopii/transforms/**/*.py",
]

[tool.ruff]
line-length = 88
target-version = "py312"
include = [
  "src/**/*.py",
  "tests/**/*.py",
]
exclude = [
  "user/**",
  "tools/vulture_whitelist.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--cov=nopii",
  "--cov-report=term-missing",
  "--cov-fail-under=40",
]

# Coverage settings removed for initial release; can be re-enabled later

[tool.uv.workspace]
members = [
    ".",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
